requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Frontend
streamlit>=1.29.0
plotly>=5.18.0

# Utilities
thefuzz[speedup]>=0.20.0
pandas>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Rate Limiting
slowapi>=0.1.9

# Logging
structlog>=24.0.0

# Caching
cachetools>=5.3.0

# Observability
sentry-sdk[fastapi]>=1.40.0
prometheus-fastapi-instrumentator>=6.1.0


# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.1.6

